requirements.txt
catboost~=1.2.5
geopandas~=1.0.1
ipython>=8.10
jupyterlab>=3.0
kedro~=0.19.8
kedro-datasets>=3.0; python_version >= "3.9"
kedro-datasets>=1.0; python_version < "3.9"
kedro-viz>=10.0.0
kedro[jupyter]
mlflow~=2.16.2
notebook
pyarrow~=17.0.0
pytest~=7.2
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
ruff~=0.1.8
scikit-learn~=1.5.1; python_version >= "3.9"
scikit-learn<=1.4.0,>=1.0; python_version < "3.9"
